+++ rest_server/controllers/context/context_meta.go
@@ -6,6 +6,6 @@ type RespMetaInfo struct {
 
 // social 리스트 정보 요청
 type SocialInfo struct {
-	SocialType int    `json:"social_type" validate:"required"`
+	SocialType int64  `json:"social_type" validate:"required"`
 	SocialName string `json:"social_name" validate:"required"`
 }
